Scope for each activity
