+++ app/src/main/java/sample/maps/injection/component/MapsActivityComponent.kt
@@ -1,0 +1,17 @@
+package sample.maps.injection.component
+
+import dagger.Subcomponent
+import sample.maps.injection.annotation.ActivityScope
+import sample.maps.injection.module.MapsActivityModule
+import sample.maps.ui.maps.MapsActivity
+
+/**
+ * Activity scoped component
+ */
+@ActivityScope
+@Subcomponent(
+        modules = arrayOf(MapsActivityModule::class)
+) interface MapsActivityComponent {
+
+    fun injectMapsActivity(mapsActivity: MapsActivity)
+}
