+++ app/src/main/java/sample/maps/injection/module/LocationListActivityModule.kt
@@ -12,19 +12,9 @@ import sample.maps.data.LocationProvider
 /**
  * Provides activity specific dependencies (or activity itself)
  */
-@Module class ActivityModule(private val activity: Activity) {
+@Module class LocationListActivityModule(private val activity: Activity) {
     @Provides
     fun providesActivity(): Activity {
         return activity
-    }
-
-    @Provides
-    fun providesRxPermissions(): RxPermissions {
-        return RxPermissions(activity)
-    }
-
-    @Provides
-    fun providesLocationProvider(@UiScheduler uiScheduler: Scheduler): LocationProvider {
-        return FusedLocationProvider(activity, uiScheduler)
     }
 }
