+++ app/src/main/java/sample/maps/ui/list/LocationListActivity.kt
@@ -6,8 +6,8 @@ import android.view.MenuItem
 import kotlinx.android.synthetic.main.location_list_activity.*
 import sample.maps.MapsApplication
 import sample.maps.R
-import sample.maps.injection.component.DaggerActivityComponent
-import sample.maps.injection.module.ActivityModule
+import sample.maps.injection.component.DaggerMapsActivityComponent
+import sample.maps.injection.module.LocationListActivityModule
 import javax.inject.Inject
 
 class LocationListActivity : AppCompatActivity() {
@@ -39,9 +39,8 @@ class LocationListActivity : AppCompatActivity() {
     }
 
     private fun initDagger() {
-        DaggerActivityComponent.builder()
-                .activityModule(ActivityModule(this))
-                .mainComponent(MapsApplication.mainComponent)
+        DaggerLocationListActivityComponent.builder()
+                .activityModule(LocationListActivityModule(this))
                 .build()
                 .injectLocationListActivity(this)
     }
