+++ app/src/main/java/sample/maps/ui/maps/MapsActivity.kt
@@ -9,8 +9,8 @@ import com.tbruyelle.rxpermissions2.RxPermissions
 import kotlinx.android.synthetic.main.maps_activity.*
 import sample.maps.MapsApplication
 import sample.maps.R
-import sample.maps.injection.component.DaggerActivityComponent
-import sample.maps.injection.module.ActivityModule
+import sample.maps.injection.component.DaggerMapsActivityComponent
+import sample.maps.injection.module.MapsActivityModule
 import javax.inject.Inject
 
 class MapsActivity : AppCompatActivity() {
@@ -41,9 +41,8 @@ class MapsActivity : AppCompatActivity() {
     }
 
     private fun initDagger() {
-        DaggerActivityComponent.builder()
-                .activityModule(ActivityModule(this))
-                .mainComponent(MapsApplication.mainComponent)
+        DaggerMapsActivityComponent.builder()
+                .activityModule(MapsActivityModule(this))
                 .build()
                 .injectMapsActivity(this)
     }
